user authentication

+++ controllers/todoController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require("express-async-handler");
 
 const Todo = require("../models/todoModel");
+const User = require("../models/userModel");
 
 // @desc      Create/Set a todo
 // @route     POST /api/todos
@@ -13,6 +14,7 @@ const setTodo = asyncHandler(async (req, res) => {
 
   const todo = await Todo.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(todo);
@@ -22,7 +24,7 @@ const setTodo = asyncHandler(async (req, res) => {
 // @route     GET /api/todos
 // @access    Private
 const getTodos = asyncHandler(async (req, res) => {
-  const todos = await Todo.find();
+  const todos = await Todo.find({ user: req.user.id });
 
   res.status(200).json(todos);
 });
@@ -36,6 +38,20 @@ const updateTodo = asyncHandler(async (req, res) => {
   if (!todo) {
     res.status(400);
     throw new Error("Todo not found!");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found!");
+  }
+
+  // Make sure the logged in user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
@@ -54,6 +70,20 @@ const deleteTodo = asyncHandler(async (req, res) => {
   if (!todo) {
     res.status(400);
     throw new Error("Todo not found!");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found!");
+  }
+
+  // Make sure the logged in user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   await todo.deleteOne();
